Rediseño de la funcion de fitness
Ahora en base a los nuevos parámetros, funciona OK, pero se puede
mejorar

+++ AG/AG/ParametrosEval.hpp
@@ -9,8 +9,8 @@ public:
 	unsigned int anchoOptimo;		// Media de ancho de sala optima
 	unsigned int altoOptimo;		// Media de alto de sala optima
 	unsigned int ciclosOptimos;		// Numero de ciclos optimo
-	unsigned int enemigosOptimos;	// Cantidad de enemigos optima
-	unsigned int cofresOptimos;		// Cantidad de cofres optima
+	double enemigosOptimos;			// % enemigos
+	double cofresOptimos;			// % cofres
 };
 
 #endif
